test(tab-button): migrate basic
