+++ core/src/components/tab-button/test/basic/tab-button.e2e.ts-snapshots/tab-button-label-ios-rtl-Mobile-Chrome-linux.png
