+++ core/src/components/tab-button/test/basic/tab-button.e2e.ts-snapshots/tab-button-label-md-rtl-Mobile-Chrome-linux.png
